requirements.txt
amqp==5.1.1
argcomplete==2.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.0.1
Beaker==1.12.1
beautifulsoup4==4.12.2
billiard==4.1.0
blivet==3.8.2
blivet-gui==2.4.2
Brlapi==0.8.5
cached-property==1.5.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.3
cupshelpers==1.0
daphne==3.0.2
dasbus==1.7
dbus-python==1.3.2
defusedxml==0.7.1
deluge==2.1.1
distro==1.8.0
dj-database-url==2.1.0
Django==4.2.4
django-allauth==0.54.0
django-cors-headers==4.2.0
django-environ==0.10.0
django-jazzmin==2.6.0
django-mathfilters==1.0.0
django-smtp-ssl==1.0
django-timezone-field==6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnf==4.18.1
docker==6.1.3
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.7
fedora-third-party==0.10
file-magic==0.4.0
fros==1.1
gbinder-python==1.1.2
GeoIP==1.3.2
gunicorn==21.2.0
humanize==4.8.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
langtable==0.0.64
libcomps==0.1.20
libdnf==0.72.0
libtorrent==2.0.9
lxml==4.9.3
Mako==1.2.3
MarkupSafe==2.1.3
nftables==0.1
numpy==1.24.4
oauthlib==3.2.2
olefile==0.46
packaging==23.1
paramiko==3.3.1
Paste==3.5.3
pexpect==4.8.0
pid==2.2.3
Pillow==10.0.0
pipdeptree==2.13.1
ply==3.11
productmd==1.37
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg==3.1.13
ptyprocess==0.7.0
pwquality==1.4.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.25.1
pyclip==0.7.0
pycparser==2.21
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.2
pyenchant==3.2.2
pygame==2.5.2
PyGObject==3.46.0
PyJWT==2.8.0
pykickstart==3.48
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyparted==3.13.0
pyrsistent==0.19.3
PySocks==1.7.1
python-augeas==1.1.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-meh==0.51
python3-openid==3.2.0
pytz==2023.3
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
rencode==1.0.6
requests==2.31.0
requests-file==1.5.1
requests-ftp==0.3.1
requests-oauthlib==1.3.1
rpds-py==0.10.6
rpm==4.19.0
selinux @ file:///builddir/build/BUILD/libselinux-3.5/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.5/python/sepolicy
service-identity==23.1.0
setools==4.4.3
setuptools==67.7.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.6.0
soupsieve==2.5
sqlparse==0.4.4
systemd-python==235
Tempita==0.5.2
texttable==1.6.4
tornado==6.3.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.3.3
whitenoise==6.5.0
zope.interface==5.5.2
